backend/requirements.txt
asgiref==3.8.1
black==24.8.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
mypy-extensions==1.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
